requirements.txt
scikit-learn==0.22.1
numpy
pandas
matplotlib
streamlit==0.76.0
Pillow
seaoborn

